+++ ChangeLog.md
@@ -3,8 +3,10 @@
 # dev 迭代中
 
 * 变更:
+
     * update: 完善shiro集成, by 大鲨鱼
-	* add: starter-ssdb by 道坤
+    * add: starter-ssdb by 道坤
+    * add: 添加NbAppEventListener,可以拦截NbApp的各种事件
 
 # 2.1.4 "平凡之路"